+++ linter/checkjs.pl
@@ -26,7 +26,6 @@ my $dirname = "htdocs/js/";
 opendir my $dir, $dirname or die "Can't open directory \"$dirname\": $!";
 while (my $entry = readdir $dir) {
     next if $entry eq "bootstrap-native.js";
-    next if $entry eq "long-press-event.js";
     next if $entry eq "i18n.js";
     next if $entry eq "apidoc.js";
     push @files, $dirname.$1 if $entry =~ /^([\w-]+\.js)$/;