+++ src/lib/config/cacertstore.c
@@ -11,7 +11,6 @@
 #include "compile_time.h"
 #include "src/lib/config/cacertstore.h"
 
-#include "src/lib/filehandler.h"
 #include "src/lib/log.h"
 #include "src/lib/sds/sds_file.h"