+++ CHANGELOG.md
@@ -12,7 +12,8 @@ This release removes the long-press event handling completely. It seems it can n
 
 - Feat: Improve playlist shuffle performance
 - Upd: Use signalfd
-- Upd: Improve websocket hanlding
+- Upd: Improve websocket handling
+- Upd: Embedded libmdpclient to 2.28 (master branch)
 - Fix: Remove long-press events #1556
 - Fix: Ignore SIGPIPE