+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This release removes the long-press event handling completely. It seems it can n
 ### Changelog
 
 - Feat: Improve playlist shuffle performance
+- Upd: Mongoose 7.23
 - Upd: Improve websocket handling
 - Upd: Embedded libmdpclient to 2.28 (master branch)
 - Fix: Remove long-press events #1556